flask-app/requirements.txt
flask
boto3
azure-storage-blob
python-dotenv
